MB-55456: History retention settings should not exceed uint64_t max
We currently allow the passing of values greater than a uint64_t max to
memcached which will get unhappy should we attempt to pass this value to
it. Prevent passing values larger than this.

In a future change we should adjust the historyRetentionSeconds to a
more appropriate max value.

Change-Id: Ifa256502d426e31f2633c5b2020ea0df7de1b12c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186391
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>