MB-59699: [BP] Disable "prevent overlapping partitions" in erl dist
Disable the new check in erlang 25 that "prevents overlapping
partitions". It is tied to the new facility of the "global" module
that gives a few features such as name registration and simple locking
across a cluster. We do not use this module, but instead have
implemented this ourselves. Perhaps this upgraded "global" module
could be used in place of our custom solution, but in the meantime
this causes problems while nodes are attempting to join together in
some situations. It has also caused some problems with remote shell
connections. Changing this setting puts this feature back to the way
it was in erlang 24 so we can decide at another time if it's worth
using the new changes from the global module.

Here are some link(s) that explains the new feature and why this is
problematic:
 - https://www.erlang.org/patches/otp-25.0
 - https://github.com/erlang/otp/pull/5611
 - https://github.com/erlang/otp/pull/5687
 - https://stackoverflow.com/q/73567169
 - https://stackoverflow.com/a/73578740

This should have originally been tagged with MB-54582, as this was
introduced with the erlang 25 upgrade.

Change-Id: Ibd810eadf4d0716e399b4d5c5f6c2c60a6b1675e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201009
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>