MB-55588/MB-55747: Fix couch_docs_fragmentation for old endpoint
This commit fixes the calculation for
/pools/default/buckets/{bucket}/stats/couch_docs_fragmentation

The calculation is now:
((couch_docs_disk_size-ep_db_history_file_size)-couch_docs_data_size)/
(couch_docs_disk_size-ep_db_history_file_size)

This additionally adds the ep_db_history_file_size stat to the bucket
stats endpoint, as we need to add the stat in order to use it in this
calculation.

For context, we originally attempted to fix this by making the
calculation in a prometheus query generated in stat_names_mappings.erl,
to avoid adding the history stat to the endpoint. This fails to
correctly aggregate the values over multiple nodes, and no way around
that was found.

Change-Id: I735ea5e55b4893e16e390d726d18c502c1e094b2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186808
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>