MB-55674: Don't use new internal tls setting for dist (add_node)
When attempting to use a new internal tls setting for dcp replication
and other tls use cases, we explicitly did not want to use this on the
distribution (mostly because it's much more challenging). This call
didn't originally seem to be a part of that process, but it clearly is
after tracing through it again. The fix is to simply use the old
options, as we did before. It can be a little deceptive, because if
you look at the call and the error, it seems as though we clean the
options in the call, but the real problem is we didn't start the
distribution with these settings in the first place, and this is where
the incompatibility lies.

Basically we call add_node with settings B when really we want to
connect with settings A. This can only be seen when using custom certs
because there is a case statment that determines whether to use these
new settings or not.

Change-Id: I01ecdaffc1b8aac6013143fec67d43fb6dcc5915
Reviewed-on: https://review.couchbase.org/c/ns_server/+/187240
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Amirbahador Biglarbegian <amir.biglarbegian@couchbase.com>