MB-56763: history_retention_collection_default() must check compat mode
ns_buckethistory_retention_collection_default() must check that the
compat mode is 7.2.0 (or greater) before returning the default history
value. This is because it is used for default values such as when a user
does not specify history on a collection or the default manifest.

Change-Id: If59fecbfb3385551a45bc6a31a0904bd7f13b021
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190586
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>