MB-57454: Replace _deleted keys with Default in dump-guts
The value of some keys in ns_config is expected to be of a specific
type in order for dump-guts to not crash. For example, the {node,_,
event_log} key should be a list, possibly containing {filename,Path}.

If a key is set to _deleted, then this assumption may fail, and cause a
crash. If we instead use Default as the value, then we get the expected
behaviour of a _deleted key essentially not being found.

Change-Id: Idb3134b3f85a45101e2de5666ff65478474c1603
Reviewed-on: https://review.couchbase.org/c/ns_server/+/192940
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>