MB-58335 [BP]: Upload cluster config on memcached restart
This is a backport of MB-55100

Prior to this change if memcached restarted (e.g. due to OOM kill) we
wouldn't upload the cluster config to memcached. As a result SDKs would
not be able to bootstrap.

This changes the terse_cluster_info_uploader to use the same mechanism
as the memcached_config_mgr to monitor memcached and restart if
memcached goes down. When the terse_cluster_info_uploader starts again
it will refresh the cluster config in memcached.

Change-Id: I6cd179a7bf43a03115b068db3887e82bd4822cbb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/195830
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>