MB-58645: Fix alerts when history retention enabled
When we check for the history_size_warning alert, we assume that there
is an ns_memcached instance for the bucket being checked.
This means that if there is a bucket with a history retention size and
time specified, then menelaus_web_alerts_srv will crash repeatedly on
every node that doesn't have that bucket.

This patch adds a check to avoid checking buckets that aren't on the
node, and adds a try ... catch incase there is still no ns_memcached
server for a bucket which has the node in its servers list.

Change-Id: I025b90ea2bbcbd789009968395c475609c10e2b8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197113
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Peter Searby <peter.searby@couchbase.com>