MB-58898: Check failover possible in quorum after config sync
It is currently possible to perform the failover "is_possible" check
on a stale version of the config before we take a quorum. If taking
the quorum takes a substantial amount of time (some number of
seconds) then the config that we performed the check on can be
substantially stale.

This issue was seen to manifest via auto-failover. A cluster was
split with a network partition such that one side had a quorum, and
the other did not. The side with quorum failed over data nodes on the
side without. The side without quorum attempted to fail over data
nodes on the side with the quorum. When the network partition was
healed the side without quorum was attempted to take the quorum to
perform a failover having already decided that one was possible. This
node remained the orchestrator, but did not realise til it was too
late (i.e. after taking the quorum and syncing the config) that nodes
had been failed over on its side of the partition and the failover
should no longer be possible. It proceeded to fail over all of the
data nodes in the cluster.

Correct this issue by checking if failover is possible after taking
the quorum and syncing the config. In future changes (in the master
branch) we can use quorum read transactions here instead.

The checks in auto-failover suffer from a similar issue and will be
addressed in followup changes.

Change-Id: If6485ea0cc86dda33ee5ea2a00f5b875c53f72fc
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199096
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>