MB-59270: [BP] Change derived stat names for backwards compatibility
This is a backport of MB-58069.

For backwards compatibility (e.g. clients of /metrics) we will report
the following stats:

sys_cpu_utilization_rate is calculated as either:
        irate(sys_cpu_cgroup_seconds_total{mode=`usage`}
  or
        sys_host_cpu_utilization_rate
depending if Couchbase is part of a cgroup.

sys_cpu_user_rate is calculated as either:
        irate(sys_cpu_cgroup_seconds_total{mode=`user`}
  or
        sys_host_cpu_user_rate
depending if Couchbase is part of a cgroup.

sys_cpu_sys_rate is calculated as either:
        irate(sys_cpu_cgroup_seconds_total{mode=`sys`}
  or
        sys_host_cpu_sys_rate
depending if Couchbase is part of a cgroup.

sys_cpu_throttled_rate is calcuated as:
        irate(sys_cpu_cgroup_seconds_total{mode=`throttled`}
if Couchbase is part of a group. Otherwise it is not reported,

sys_cpu_burstd_rate is calcuated as:
        irate(sys_cpu_cgroup_seconds_total{mode=`burst`}
if Couchbase is part of a group. Otherwise it is not reported,

Change-Id: Id433517da545733219f258fca59b4cbf607ca678
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200418
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>