MB-58197: [BP] Add err msg for timeout of node during pools default POST
This is a backport of MB-54380.

We currently don't handle a timeout gracefully and instead hit a
bad match and return a HTTP 500. We probably don't want to change
the behaviour here, it would be a bad idea to allow a cluster quota
to be set that was too high for a node that may (temporarily) not be
communicable. Add an explicit error message and return a 500 status:

$ curl -v -X POST localhost:9000/pools/default -u Administrator:asdasd -d memoryQuota=1026 -w "\n"
< HTTP/1.1 500 Internal Server Error
< Cache-Control: no-cache,no-store,must-revalidate
< Content-Length: 61
< Content-Type: text/plain
< Date: Fri, 03 Mar 2023 14:20:27 GMT
< Expires: Thu, 01 Jan 1970 00:00:00 GMT
< Pragma: no-cache
< Server: Couchbase Server
< X-Content-Type-Options: nosniff
< X-Frame-Options: DENY
< X-Permitted-Cross-Domain-Policies: none
< X-XSS-Protection: 1; mode=block
<
* Connection #0 to host localhost left intact
Did not receive response in time from nodes ['n_2@127.0.0.1']

Change-Id: I1566f67718e5eae3705435dc802cbfbb7a1e5561
Reviewed-on: https://review.couchbase.org/c/ns_server/+/200280
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>