MB-58898: Remove leader activity in auto_failover
It is problematic to run auto_failover within a leader activity.
Whilst possible, we would need to run the activity on a non-default
domain to avoid domain conflict with any other leader activity. The
orchestrator inherently serializes leader activities and we need to
continue to run leader activities on the default domain via the
orchestrator to avoid having to write conflict handling.

The goal of this change was to avoid passing auto failover checks
on a stale config and continuing to auto-failover nodes that were not
valid to fail over. This change does so to the same affect but by
re-validating the safety of auto_failover within the existing
leader activity. This removes the needs to enter a leader activity
within auto failover before we attempt the failover. In future changes
we will aim to use a single transaction in failover and this solution
makes that easier.

Change-Id: I3af7ae9052a1c6300656b605627f73160ebbd970
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201590
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>