MB-59455 do check for cluster uuid prior to the authentication
The reason for this change is that cbauth client would like to
make sure that all http checks are made against the correct cluster.

Currently it is done by adding uuid parameter to the http request,
so the request returns 404 if the specified value doesn't
match the actual cluster uuid.

Unfortunately this check will be performed for /_cbauth endpoint
only if the authentication succeeds, so there's no way to check
if the negative result was obtained with correct cluster uuid.

Change-Id: I1e7053ebd4716f7e8f90f4224a0bcd96229585fa
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201913
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Restriction Checker