MB-59104: Fix issue with self eject http response
Currently when a node is self ejected from the cluster via REST, it
will cause ns_cluster:leave_body() to be executed while the REST
request is being handled. As part of the leave, the node will
terminate the ns_server_nodes_sup so node data can be cleaned up and
the ns_server_nodes_sup restated after. All this happens while the
client is still waiting for a response to the REST request. However
the client will not get a response because the web server is shut
down as part of the process of shutting down the nodes supervisor.

This fix ensures the client is able to receive a response by trapping
the exit, responding to client, and then exiting.

Change-Id: I9fe9f3a1fddec77ff7d70d67b35465c941b169b0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/228702
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>