MB-61986 [BP]: Rebalance progress metrics
This is a backport of MB-57610

At each scrape, there will be one metric for each service, labelled as
"stage". Once the rebalance starts, the progress will be set to 0 for
all stages. Once the rebalance of that service completes, the progress
for that stage will be set to 1. If the rebalance terminates early, or
is stopped, the progress will be left at the last value reached, until
the next rebalance starts and it gets set back to 0.

Change-Id: Iaaf36f31d7b30aa63bf400619c89fe65fbfe6d08
Reviewed-on: https://review.couchbase.org/c/ns_server/+/212133
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>