[BP] MB-60568 Use 'online' processor count
Prior to this change we used the number of logical_processors to
determine the cores for the node. This is number of logical processors
configured in the system. In the case of a VM running on a hypervisor
the number of processors allocated to the VM may be a subset of the
overall number. So we now will use the number of
logical_processors_online.

More info is available at:

https: //www.erlang.org/doc/man/erlang#system_info_logical_processors
Change-Id: I3c3d4403fb3382dcc9abc1ad420a64619c23faf3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/205058
Well-Formed: Restriction Checker
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>