MB-60634: prometheus_cfg terminate loglevel change
Change the loglevel for the prometheus_cfg gen_server terminate hook s/t
if the "Reason" is "shutdown" we should instead log it as "info" level
because it is normal when restarting the service. Leaves all other
"Reasons" to log at error level so we still catch any other reasons for
shutting down.

Change-Id: Ifa9aa1a1806a44d213c984e863c0bff339787e8a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/207116
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>