menelaus_web_guardrails: assert ns_config call was expected
The original test attempted to check whether ns_config was correctly
updated, but it was missing an assert and wouldn't actually have been
passing because it didn't allow for the order of proplists to change.

By explicitly testing this, we ensure that any future changes to what
ends up in ns_config are actually tested.

Change-Id: I5f8367a71c61c98d491817f87423e06d5f36ad97
Reviewed-on: https://review.couchbase.org/c/ns_server/+/207497
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>