MB-61259 Remove ns_tick_agent and pubsub race
Prior to this change the ns_tick_agent would obtain the mb master node
and then subscribe to leader_events.  If the master determination was
made between the two calls the ns_tick_agent would not receive the event
and would not have the mb master node.

Change-Id: I4b3d2a24f228b6c324eb50f9b502119a16551337
Reviewed-on: https://review.couchbase.org/c/ns_server/+/207726
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>