MB-61405: Report correct stat name for disks like "pxd/etc"
The disk name and stat name were handled using code reused from
handling proc stats, which didn't correctly handle "/" in the disk
name. This meant that the end of the disk path got inserted in the
"sys_disk_<stat name>" metric reported to prometheus.

The proc stat handling code adds a "/", which it then splits back out
later. It's likely that the reason for this was for backwards
compatiblity, which is not a concern for the disk stats since they were
added post 7.0.

Change-Id: Ia0eb863f17d4938f8a048272d35fba224ac0eb22
Reviewed-on: https://review.couchbase.org/c/ns_server/+/208099
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Restriction Checker