MB-61648: Clearer msg when specifying IP address for cbcollect
When an IP address was specified for the uploadhost the default scheme
is secure. The result is a tls handshake error. Prior to this change the
message provided by the UI was confusing. This attempts to make it
clearer.

This change is dependent on:
  https://review.couchbase.org/c/couchdb/+/209294

Change-Id: I13249a25cb08196e74f93871fa8fb8a7426c4f15
Reviewed-on: https://review.couchbase.org/c/ns_server/+/209168
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>