MB-61276: Properly close audit stream on 'EXIT'
Currently, if you are streaming chunked responses from menelaus_web, and
an error occurs, we were only catching 'EXIT' messages of type 'normal'
instead of any other type of 'EXIT' message. This caused us to not
properly close the chunked stream (by writing an empty chunk) and the
reader on the other side of the connection, will likely read the \r\n
carriage return that are outside the octal range required for the "size"
field of stream encoding. That's why you would see "invalid byte in
chunk length" because it basically never gets the end-chunk message, but
instead gets something outside the proper range (\r\n) and throws this
error.

This fixes it by simply catching all 'EXIT' messages, closing the
chunked-stream and then re-raising the message with exit(Type) so the
crash can continue to bubble up the stack.

We also log a warning stating that an abnormal 'EXIT' message was
received.

Also includes a cluster_test that will listen on a streaming endpoint,
remove that node, and verify that we never get the infamous "invalid
byte in chunk length" error but instead get the EOF variant we expect.
It uses the actual golang library code to do this s/t we can test the
exact situation from the ticket.

Change-Id: Ia5a6716fa0b59d807da424b959029473614b1119
Reviewed-on: https://review.couchbase.org/c/ns_server/+/209302
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>