MB-58691: Overwrite passed cluster with result from testset
This allow us to change the cluster variable of a testset and have the
modification manifest for the testrunner to correctly tear down the
correct cluster. This will be used in a future change to test node
remapping which tears down and recreates the original cluster at the
start and end of the test respectively.

Change-Id: I1ce5b15f03ff3ad9542fb76db4ed57705de97480
Reviewed-on: https://review.couchbase.org/c/ns_server/+/209394
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>