MB-61201 Disable indexing guardrails
Indexing guardrails are no longer targeted for 7.6.2, so they should be
disabled.

Ideally we would just revert the patches, but since there's quite a
large number of them, and they're mixed in with some code/test
improvements, it's more work than it's worth to revert them.
Particularly because we would be un-reverting them immediately in
master (so indexing can test their patches against ours).

Instead, we are just removing the api and configuration, so the index
guardrails will not be enforced, and no extra config keys will be sent
to indexing.

Change-Id: I53e22231672aa84ee3b311fac607eb7978fab1ea
Reviewed-on: https://review.couchbase.org/c/ns_server/+/209569
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Restriction Checker