MB-62036: [BP] Add cm_is_balanced stat
This backports:
MB-61810: Add cm_is_balanced stat

The cm_is_balanced stat is a gauge returning 0 (false) or 1 (true) on
whether or not the cluster is balanced. It is the same value as returned
by "isBalanced" from terseClusterInfo.

The stat is populated into the ns_server_stats ets table by a separate
proess so as to not block other work by what could be a time consuming
operation. As it is a fairly heavyweight stat it is populated less
frequently (every 30 seconds) and only reported by the orchestrator
node.

Change-Id: I8a3719b1ab0d07750a18ae868ef0f8f572887a46
Reviewed-on: https://review.couchbase.org/c/ns_server/+/210472
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>